fetch function code refactor
