+++ mews-task-movies/src/pages/homePage/homePage.tsx
@@ -1,30 +1,14 @@
 import { SetStateAction, useEffect, useState } from 'react';
 import SearchForm from '../../components/searchForm/searchForm';
 import MovieList from '../../components/movieList/movieList';
+import { getMovies } from '../../data/getMovies';
 
 export default function HomePage() {
   const [movies, setMovies] = useState([]);
   const [search, setSearch] = useState('');
 
   useEffect(() => {
-    const getMovies = async () => {
-      const options = {
-        method: 'GET',
-        headers: {
-          accept: 'application/json',
-        },
-      };
-      const response = await fetch(
-        `https://api.themoviedb.org/3/search/movie?query=${search}&api_key=03b8572954325680265531140190fd2a`,
-        options,
-      );
-      const moviesData = await response.json();
-
-      console.log('moviesData', moviesData);
-      setMovies(moviesData.results);
-      console.log('moviesData.result', moviesData.results);
-    };
-    getMovies();
+    getMovies(search, setMovies);
   }, [search]);
 
   const handleSearch = (e: { target: { value: SetStateAction<string> } }) => {
